feat(i18n): replace static uploadTime string with live i18n relative time
Remove the hardcoded English formatUploadTime() from trendAnalysisService
and the VideoData.uploadTime field. Add a shared formatTimeAgo(ts, t)
utility in formatters.ts that uses the existing timeAgo.* i18n keys.

VideoCard, HighlightVideoCard, VideoListTable and FavoriteRow now call
formatTimeAgo at render time, so the timestamp label reflects the active
locale (e.g. German "vor 3 Std." instead of static "3h ago").
FavoriteRow's local duplicate formatTimeAgo is also removed.
