+++ src/shared/lib/formatters.ts
@@ -2,6 +2,7 @@
  * Formatting utilities
  */
 
+import type { TFunction } from "i18next";
 import { getLocale } from "./locale";
 
 /**
@@ -54,4 +55,31 @@ export function formatDuration(ms: number): string {
  */
 export function formatPercentage(value: number, decimals = 0): string {
   return `${value.toFixed(decimals)}%`;
+}
+
+/**
+ * Format a past timestamp as a localized relative time string.
+ * Uses the timeAgo.* i18n keys present in all supported locales.
+ *
+ * @param timestamp - Unix timestamp in milliseconds
+ * @param t         - i18next TFunction from useTranslation()
+ */
+export function formatTimeAgo(timestamp: number, t: TFunction): string {
+  const diffMs = Date.now() - timestamp;
+  if (diffMs < 0) return t("timeAgo.justNow");
+  const sec = Math.floor(diffMs / 1000);
+  if (sec < 10) return t("timeAgo.justNow");
+  if (sec < 60) return t("timeAgo.seconds", { count: sec });
+  const min = Math.floor(sec / 60);
+  if (min < 60) return t("timeAgo.minutes", { count: min });
+  const hrs = Math.floor(min / 60);
+  if (hrs < 24) return t("timeAgo.hours", { count: hrs });
+  const days = Math.floor(hrs / 24);
+  if (days < 7) return t("timeAgo.days", { count: days });
+  const weeks = Math.floor(days / 7);
+  if (weeks < 5) return t("timeAgo.weeks", { count: weeks });
+  const months = Math.floor(days / 30);
+  if (months < 12) return t("timeAgo.months", { count: months });
+  const years = Math.floor(days / 365);
+  return t("timeAgo.years", { count: years });
 }
